packet structure and command structure (and related lines of code) altered

+++ client_ftp.h
@@ -8,7 +8,7 @@
 
 struct command
 {
-	int id;
+	short int id;
 	int npaths;
 	char** paths;
 };
